Fix bug where sidebar pfp don't update for new usr
There was a bug where if a user made an account, logged in, and changed their PFP, the pfp in the sidebar wouldn't update. Now that bug is fixed.

+++ Project/app/src/main/java/com/example/cmpt362project/ui/settings/UserProfileActivity.kt
@@ -207,12 +207,18 @@ class UserProfileActivity : AppCompatActivity() {
 
                 // Write the new profile picture path to the user's info
                 userReference.setValue(newImgPath).addOnSuccessListener {
-                    println("$printIdentifier: Uploaded $randomUUID to database")
+                    println("$printIdentifier: Uploaded $newImgPath to database")
 
                     // Delete the old profile picture from Storage, tell sidebar to update PFP
-                    val oldImgRef = storage.child(oldImgPath.value as String)
-                    oldImgRef.delete().addOnSuccessListener {
-                        println("$printIdentifier: Deleted ${oldImgPath.value} from database")
+                    // Do not delete the old PFP if it's the default one
+                    val pathToDelete = oldImgPath.value as String
+                    if (pathToDelete != getString(R.string.default_pfp_path)) {
+                        val oldImgRef = storage.child(oldImgPath.value as String)
+                        oldImgRef.delete().addOnSuccessListener {
+                            println("$printIdentifier: Deleted ${oldImgPath.value} from database")
+                            updateProfilePicSharedPref()
+                        }
+                    } else {
                         updateProfilePicSharedPref()
                     }
                 }
